requirements.txt
pytesseract~=0.3.13
Pillow~=11.1.0
python-Levenshtein
pdf2image~=1.17.0
augraphy~=8.2.6

opencv-python~=4.11.0.86